+++ ptf/Makefile
@@ -1,8 +1,6 @@
-PTF_CMD = sudo -E ./ptf_runner.py --device stratum-bmv2 --port-map port_map.veth.json --ptf-dir tests --cpu-port 255 --device-id 1
+PTF_CMD = sudo -E python2.7 runner.py --device stratum-bmv2 --port-map port_map.json --ptf-dir tests --cpu-port 255 --device-id 1
 P4C_SRC_DIR = ../p4src
 P4C_OUT = ../p4src/build
-
-all: test
 
 test: p4src
 ifndef TEST
@@ -17,3 +15,7 @@ p4src:
 
 %:
 	make test TEST=$@
+
+clean:
+	rm -rf ptf.log
+	rm -rf ptf.pcap
